CustomObjectRegistry instead. The CustomObjectCollection then uses that.
