+++ TimberAPI/CustomObjectSystem/CustomObjectCollection.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Timberborn.AssetSystem;
+using Timberborn.EntitySystem;
+using UnityEngine;
+
+namespace TimberbornAPI.CustomObjectSystem
+{
+    public class CustomObjectCollection : IObjectCollection
+	{
+
+		public IEnumerable<Object> GetObjects()
+		{
+			return TimberAPI.CustomObjectRegistry.GetAllGameObjects();
+		}
+	}
+}
