+++ TimberAPI/TimberAPIPlugin.cs
@@ -4,9 +4,9 @@ using HarmonyLib;
 using Timberborn.ToolPanelSystem;
 using TimberbornAPI.AssetLoaderSystem.AssetSystem;
 using TimberbornAPI.Common;
+using TimberbornAPI.CustomObjectSystem;
 using TimberbornAPI.DependencySystem;
 using TimberbornAPI.EntityActionSystem;
-using TimberbornAPI.ObjectCollectionSystem;
 using TimberbornAPI.UIBuilderSystem;
 using UnityEngine;
