+++ TimberAPIExample/Examples/CustomObjectRegistryExample/CustomObjectExampleConfigurator.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Bindito.Core;
+
+namespace TimberAPIExample.Examples.CustomObjectRegistryExample
+{
+    public class CustomObjectExampleConfigurator : IConfigurator
+    {
+        public void Configure(IContainerDefinition containerDefinition)
+        {
+            containerDefinition.Bind<CustomObjectAddExample>().AsSingleton();
+        }
+    }
+}
